+++ packages/core/src/components/expandable/expandable.interfaces.ts
@@ -1,0 +1,8 @@
+export interface ExpandableInterface {
+  getAnimeInstance(): Promise<anime.AnimeInstance | undefined>;
+}
+
+export interface ExpandableToggleEvent {
+  originalEvent?: MouseEvent | KeyboardEvent;
+  open: boolean;
+}