+++ packages/dso-toolkit/src/components/expandable/expandable.args.ts
@@ -1,0 +1,37 @@
+import { HandlerFunction } from "@storybook/addon-actions";
+import { ArgTypes, noControl } from "../../storybook/index.js";
+import { Expandable } from "./expandable.models.js";
+
+export interface ExpandableArgs {
+  open: boolean;
+  enableAnimation: boolean;
+  animationOffset: number;
+  dsoToggle: HandlerFunction;
+}
+
+export const expandableArgTypes: ArgTypes<ExpandableArgs> = {
+  open: {
+    type: "boolean",
+  },
+  enableAnimation: {
+    type: "boolean",
+  },
+  animationOffset: {
+    ...noControl,
+  },
+  dsoToggle: {
+    ...noControl,
+    action: "dsoToggle",
+  },
+};
+
+export function expandableArgsMapper<TemplateFnReturnType>(
+  a: ExpandableArgs,
+  content: TemplateFnReturnType
+): Expandable<TemplateFnReturnType> {
+  return {
+    ...a,
+    content,
+    dsoToggle: (e) => a.dsoToggle(e.detail),
+  };
+}