+++ packages/react/src/components/expandable/expandable.content.tsx
@@ -1,0 +1,10 @@
+import { html, TemplateResult } from "lit-html";
+
+import React from "react";
+
+export const expandableContent: JSX.Element = (
+  <div className="dso-rich-content">
+    <h3>Expandable</h3>
+    <span>Dit is een expandable</span>
+  </div>
+);