+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 * Accordion: Neutral variant ([#2099](https://github.com/dso-toolkit/dso-toolkit/issues/2099))
 * Actielijst ([#2098](https://github.com/dso-toolkit/dso-toolkit/issues/2098))
+* Webcomponent `<dso-expandable>` ([#2034](https://github.com/dso-toolkit/dso-toolkit/issues/2034))
 
 ### Changed
 * **all:** Package updates ([#2118](https://github.com/dso-toolkit/dso-toolkit/issues/2118))